Test askToMove
Create a new test file that checks the correct output is sent to cout,
depending on the value of playerTurn.

In order to test this, pass a reference to an ostream variable into the
function with a default value of std::cout. This will allow an
ostringstream variable to be passed into the function call in the test
which can then be checked to see if it contains the correct string
message.