Test playAgain
Create a new test file that checks playAgain returns true or false,
depending on the input given by the user. Use an istringstream object to
mimic cin.

Also, check that it prints out a message letting the player know when they
enter invalid input. Do this by passing in another boolean variable that
is set to false by default. When true, break out of the while loop so that
the test does not run forever.