Test fillCell
Create a test file for the fillCell function. In order to test that the
correct functions are called, make two new parameters and set their default
values to fillCellX and fillCellO so that a mock function can be passed
through in the tests. Then, check the correct commands have been carried
out on a string variable called result.